requirements.txt
pynput
Pillow
pywin32
cryptography
pyperclip
sounddevice
scipy
numpy
opencv-python
matplotlib
